Game links against librw
Still a lot more to go...

+++ src/SB/Core/x/xModel.h
@@ -152,7 +152,7 @@ void xModelResetMaterial(xModelInstance* model);
 void xModel_SceneEnter(RpWorld* world);
 void xModel_SceneExit(RpWorld* world);
 xSphere* xModelGetLocalSBound(xModelInstance* model);
-void xModelGetBoneMat(xMat4x3& mat, const xModelInstance& model, size_t index);
+void xModelGetBoneMat(xMat4x3& mat, const xModelInstance& model, u32 index);
 void xModelInstanceUpgradeBrotherShared(xModelInstance* inst, U32 flags);
 xVec3 xModelGetBoneLocation(const xModelInstance& model, u32 index);
